fix(bun): resolve preload modules when cwd differs from project root
Lazy-load Babel via createRequire in the solid preload, extract a
logger-free sqlite-vec bootstrap for preloads, and drop duplicate test
preloads already covered by the top-level bunfig entries.
