+++ bunfig.toml
@@ -7,7 +7,5 @@ minimumReleaseAge = 604800
 [test]
 preload = [
     "./src/utils/bun/preload-test-process-exit.ts",
-    "./src/utils/bun/preload-solid-scoped.ts",
-    "./src/utils/search/stores/sqlite-vec-preload.ts",
     "./src/indexer/lib/test-cleanup-preload.ts",
 ]
