+++ package.json
@@ -89,6 +89,7 @@
         "@ast-grep/napi": "^0.42.0",
         "@babel/core": "^7.28.6",
         "@babel/preset-typescript": "^7.28.5",
+        "babel-preset-solid": "^1.9.10",
         "@clack/prompts": "^1.0.0",
         "@crawlee/utils": "^3.16.0",
         "@genesiscz/darwinkit": "0.7.5",
