+++ src/utils/search/stores/sqlite-vec-loader.ts
@@ -1,51 +1,15 @@
-import { Database } from "bun:sqlite";
-import { existsSync } from "node:fs";
+import type { Database } from "bun:sqlite";
 import { logger } from "@app/logger";
+import {
+    ensureExtensionCapableSQLiteCore,
+    getSqliteVecLoaderState,
+    HOMEBREW_SQLITE_PATHS,
+    resetSqliteVecLoaderState,
+} from "./sqlite-vec-bootstrap";
 
-/**
- * `Database.setCustomSQLite()` is a *process-global, one-shot* native call.
- * The guard that tracks "have we already attempted it?" must therefore live
- * at process scope too, NOT module scope.
- *
- * This module is loaded TWICE in a normal `tools` invocation: once via the
- * launcher's `--preload <absolute path>` and once via bunfig.toml's relative
- * `./src/...` preload entry. Bun keys its module cache by specifier, so the
- * absolute and relative paths resolve to two distinct module instances. A
- * module-level `let` guard is `false` in each, so the second instance called
- * `setCustomSQLite()` again — the native side throws "SQLite already loaded"
- * and we logged a spurious WARN (15-27/day) even though the first call had
- * already swapped successfully and sqlite-vec was working fine. Hanging the
- * state off `globalThis` makes both instances share one guard.
- */
-interface SqliteVecGlobalState {
-    extensionAvailable: boolean | null;
-    customSqliteAttempted: boolean;
-    homebrewDylibFound: boolean;
-}
+export { HOMEBREW_SQLITE_PATHS };
 
-const STATE_KEY = Symbol.for("genesis-tools.sqlite-vec.loader-state");
-
-function state(): SqliteVecGlobalState {
-    const g = globalThis as typeof globalThis & {
-        [STATE_KEY]?: SqliteVecGlobalState;
-    };
-
-    if (!g[STATE_KEY]) {
-        g[STATE_KEY] = {
-            extensionAvailable: null,
-            customSqliteAttempted: false,
-            homebrewDylibFound: false,
-        };
-    }
-
-    return g[STATE_KEY];
-}
-
-/** Known paths for Homebrew sqlite3 with extension support (arm64 first) */
-const HOMEBREW_SQLITE_PATHS = [
-    "/opt/homebrew/opt/sqlite3/lib/libsqlite3.dylib",
-    "/usr/local/opt/sqlite3/lib/libsqlite3.dylib",
-];
+const state = getSqliteVecLoaderState;
 
 /**
  * Try to swap bun:sqlite to a build that supports extension loading.
@@ -56,63 +20,7 @@ const HOMEBREW_SQLITE_PATHS = [
  * your program entry point if you plan to use sqlite-vec.
  */
 export function ensureExtensionCapableSQLite(): void {
-    const s = state();
-
-    if (s.customSqliteAttempted) {
-        return;
-    }
-
-    s.customSqliteAttempted = true;
-
-    if (process.platform !== "darwin") {
-        // Linux and Windows ship Bun with extension-capable SQLite already.
-        s.homebrewDylibFound = true;
-        return;
-    }
-
-    for (const libPath of HOMEBREW_SQLITE_PATHS) {
-        if (existsSync(libPath)) {
-            try {
-                const swapped = Database.setCustomSQLite(libPath);
-                if (swapped === false) {
-                    logger.warn(
-                        `[sqlite-vec] setCustomSQLite(${libPath}) returned false - a Database may have been ` +
-                            "created before the preload ran; sqlite-vec will be unavailable."
-                    );
-                    return;
-                }
-
-                s.homebrewDylibFound = true;
-                logger.debug(`[sqlite-vec] swapped in extension-capable SQLite: ${libPath}`);
-            } catch (err) {
-                const message = err instanceof Error ? err.message : String(err);
-
-                // "SQLite already loaded" is the benign double-attempt: another
-                // module instance / duplicate preload already ran the one-shot
-                // swap in THIS process (the global guard makes that a no-op, but
-                // a path that bypasses it still lands here). If the earlier
-                // attempt succeeded, sqlite-vec works fine; if a Database loaded
-                // SQLite before any swap, loadSqliteVec() reports the real,
-                // actionable failure with its own warn. Either way THIS throw
-                // is not actionable — debug, not warn, so it stops being noise.
-                if (message.includes("SQLite already loaded")) {
-                    logger.debug(
-                        `[sqlite-vec] setCustomSQLite(${libPath}) skipped - SQLite already loaded ` +
-                            "in this process (swap attempted by an earlier module instance / preload)."
-                    );
-
-                    return;
-                }
-
-                logger.warn(
-                    `[sqlite-vec] setCustomSQLite(${libPath}) failed - sqlite-vec will be unavailable. ` +
-                        `Cause: ${message}`
-                );
-            }
-
-            return;
-        }
-    }
+    ensureExtensionCapableSQLiteCore(logger);
 }
 
 /**
@@ -165,9 +73,6 @@ export function loadSqliteVec(db: Database): boolean {
         return false;
     }
 
-    // ensureExtensionCapableSQLite() should already have been called before
-    // any Database was created. Call it here as a safety net, but it may be
-    // too late if a Database instance already exists.
     if (!s.customSqliteAttempted) {
         ensureExtensionCapableSQLite();
     }
@@ -214,9 +119,5 @@ export function isSqliteVecAvailable(): boolean {
  * one-shot swap and availability is re-probed.
  */
 export function resetSqliteVecState(): void {
-    const g = globalThis as typeof globalThis & {
-        [STATE_KEY]?: SqliteVecGlobalState;
-    };
-
-    delete g[STATE_KEY];
+    resetSqliteVecLoaderState();
 }
